Fix gcm_setup to not crush tmpl files
I did not understand how this really worked and @bmauer found out it was
working for @mathomp4 because @mathomp4 "fixed it".

In truth, the code was crushing installed files in `install/etc`. Now it
is a bit more careful by copying from `install/etc` and then doing all
manipulations in `install/bin`. Should probably be doing this in a
`mktemp` directory, but seems to work for now. At least, @bmauer could
run