Fix option: _dual_ocean_ since it applies to MOM5
The option to run coupled model in with the `DUAL_OCEAN:1` only works with MOM5 ocean model. 
It does not and can not work with MOM6 ocean model, [see for more details](https://github.com/GEOS-ESM/GEOSgcm_GridComp/issues/322).

Fixing the `AGCM.rc.tmpl` is essential to avoid any confusion about this option and its functionality.