v11: Updates for Intel MPI at NAS
Testing is currently underway to perhaps move to Intel MPI at NAS due to issues running v12 with Ops emissions. With MPT, it fails with:
```
MPT ERROR: Cannot create more than 2048 RMA windows.
```
and per NAS support, that is a hard limit compiled into MPT.

So, Intel MPI. And in doing so, it seems the flags we use at NCCS are good at NAS (in testing, so far).

Note: While this is not (currently) happening in v11, I want to capture
the changes in case it does. Plus, we might move to Intel MPI for
simplicity.